refactor(standard): error handling
