+++ standard/error.lua
@@ -6,16 +6,18 @@
 -- Please see https://github.com/Liquipedia/Lua-Modules to contribute
 --
 
----@class error
----@field childErrors? error[]
----@field header string?
----@field innerError any
----@field message string
----@field originalErrors? error[]
----@field stacks? string[]
----@field is_a? function
-
+local Array = require('Module:Array')
 local Class = require('Module:Class')
+local Json = require('Module:Json')
+local Page = require('Module:Page')
+local String = require('Module:StringUtils')
+local Table = require('Module:Table')
+
+local FILTERED_ERROR_STACK_ITEMS = {
+	'^Module:ResultOrError:%d+: in function <Module:ResultOrError:%d+>$',
+	'^%[C%]: in function \'xpcall\'$',
+	'^Module:ResultOrError:%d+: in function \'try\'$',
+}
 
 --[[
 A minimal error class, whose purpose is to allow additional fields to be
@@ -47,9 +49,15 @@ preamble-like text here to give some context to the error.
 error.noStack: Disables the stack trace
 
 ]]
----@class Error
+---@class Error: BaseClass
 ---@operator call(string|table|nil|any):Error
----@field message string?
+---@field message string
+---@field childErrors? Error[]
+---@field header string?
+---@field innerError any
+---@field originalErrors? Error[]
+---@field stacks? string[]
+---@field is_a? function
 local Error = Class.new(function(self, any)
 	-- Normalize the various ways an error can be thrown
 	if type(any) == 'string' then
@@ -67,7 +75,7 @@ local Error = Class.new(function(self, any)
 	self.message = self.message or 'Unknown error'
 end)
 
----@param error error
+---@param error Error
 ---@return boolean
 function Error.isError(error)
 	return type(error) == 'table'
@@ -78,6 +86,65 @@ end
 
 function Error:__tostring()
 	return self.message
+end
+
+---Builds a JSON string for use by `liquipedia.customLuaErrors` JS module with `error()`.
+---@return string
+function Error:getErrorJson()
+	local stackTrace = {}
+
+	local processStackFrame = function(frame, frameIndex)
+		if frameIndex == 1 and frame == '[C]: ?' then
+			return
+		end
+
+		local stackEntry = {content = frame}
+		local frameSplit = mw.text.split(frame, ':', true)
+		if (frameSplit[1] == '[C]' or frameSplit[1] == '(tail call)') then
+			stackEntry.prefix = frameSplit[1]
+			stackEntry.content = mw.text.trim(table.concat(frameSplit, ':', 2))
+		elseif frameSplit[1]:sub(1, 3) == 'mw.' then
+			stackEntry.prefix = table.concat(frameSplit, ':', 1, 2)
+			stackEntry.content =  table.concat(frameSplit, ':', 3)
+		elseif frameSplit[1] == 'Module' then
+			local wiki = not Page.exists(table.concat(frameSplit, ':', 1, 2)) and 'commons'
+				or mw.text.split(mw.title.getCurrentTitle():canonicalUrl(), '/', true)[4] or 'commons'
+			stackEntry.link = {wiki = wiki, title = table.concat(frameSplit, ':', 1, 2), ln = frameSplit[3]}
+			stackEntry.prefix = table.concat(frameSplit, ':', 1, 3)
+			stackEntry.content = table.concat(frameSplit, ':', 4)
+		end
+
+		table.insert(stackTrace, stackEntry)
+	end
+
+	Array.forEach(self.stacks, function(stack)
+		local stackFrames = mw.text.split(stack, '\n')
+		stackFrames = Array.filter(
+			Array.map(
+				Array.sub(stackFrames, 2, #stackFrames),
+				function(frame) return String.trim(frame) end
+			),
+			function(frame) return not Table.any(FILTERED_ERROR_STACK_ITEMS, function(_, filter)
+				return string.find(frame, filter) ~= nil
+			end) end
+		)
+		Array.forEach(stackFrames, processStackFrame)
+	end)
+
+	local errorSplit = mw.text.split(self.message, ':', true)
+	local errorText
+	if #errorSplit == 4 then
+		errorText = string.format('Lua error in %s:%s at line %s:%s.', unpack(errorSplit))
+	elseif #errorSplit > 4 then
+		errorText = string.format('Lua error in %s:%s at line %s:%s', unpack(Array.sub(errorSplit, 1, 4)))
+		errorText = errorText .. ':' .. table.concat(Array.sub(errorSplit, 5), ':') .. '.'
+	else
+		errorText = string.format('Lua error: %s.', self.message)
+	end
+	return Json.stringify({
+			errorShort = errorText,
+			stackTrace = stackTrace,
+		}, {asArray = true})
 end
 
 return Error
